requirements.txt
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
certifi==2015.11.20.1
dnspython==1.12.0
enum34==1.1.2
msgpack-python==0.4.7
MySQL-python==1.2.5
pycrypto==2.6.1
python-etcd==0.4.3
requests==2.9.1
singledispatch==3.4.0.3
six==1.10.0
tornado==4.3
torndb==0.3
urllib3==1.14
wheel==0.24.0
